requirements.txt
python-dotenv==1.0.0
python-telegram==0.18.0
telegram-text==0.1.2